Fix issue #17 - Time zones offset bug
Prior to this change the calculation of the average hourly volume
for the last hour was relying on the time stamp from the candles
as provided by the server and the local time. Because of daylight
savings, the offset could change, and break the calculation.

This change calculates the time interval using loca times only,
thus not relying on any server side provided times.