Address div by zero problems in candlesticks.py
Prior to this change, the function candlesticks.GetAverageVolume() did
not check if any of the arguments made sense, like the time frame over
which to calculate the average. Passing zero or a negative number was
OK, which could result in div by zero.

Checks are put in place, and if invalid inputs are detected, the
function will return -1. Invalid inputs are logged.

Fixes Issue #14.