Fix #6, fix #11 and improve the code on many fronts
Prior to this change, the market class would do some checks that were
more properly performed in the strategy class that required them.
There was also a bug that would trigger the Pump N' Dump strategy when
not necessary. To address this the code had to be improved on various
fronts.

All the checks for a strategy are now performed inside the strategy
class itself. However, some derived market properties are still
computed within the market class. To address the false positive in
the Pump N' Dump strategy additional checks have been added.