Improve logs of fake trades
Prior to this change, when a fake order was placed, the effective rate
at which the (fake) order was executed was hard coded. This made logging
more difficult as it was necessary to distinguish a fake from a real
trade. Now it's possible to use a single logging interface to dump
any required data, and in a fake trade this will just dump whatever the
rate submitted for the order was.

The order returns the same rate submitted for the (fake) trade.