Introduce container to hold all order details
Prior to this change, order details were passed with one specific exchange
in mind, Bittrex. At leas one of the paramteres was also essentially hard
coded.

This commit introduces an abstract container that holds all the
information required for a specific order. The container,
an OrderDetails class, also implements some sanity checks on the order
details and raises various exceptions if problems are encounterd.

Part of work required for #13 and #12.