Enhance volume estimates for current candle
Prior to this change, the only available measure was the volume that was
received from the exchange. A new metric is now available, the estimated
volume for the full tick.

The volume for the full tick is extrapolated simply by assuming a
constant gradient, i.e. assume that the volume experienced during the
current time interval is proportional to the current time fraction
relative to the full tick time interval.