Kill bug in all candles update
Prior to this change, whenver all candles were updated, the most recent
but not current candle would get replaced by the first update to the
current candle. Over time this lead to extremely low calculated daily
volumes with equally low standard devations. In short, a completely
unlreliable set of candles for the previous day.

One example of problems with this was a Pump 'n Dump would get detected
for no reason.

The order of candle updating has been changed.