Allow querying reserved balances and enhance logs
Prior to this change, you could not query directly the reserved balance
for a certain coin. (On Bittrex a balance gets reserved if there's a
trade involved, like an open order.) This allows to check for open
orders indirectly, without making an API call.

Additionally, some logging has been cleaned and refined.

A new method has been introduced that calculates the reserved balance
as a difference between total balance and available balance.