Remove the `market` object dependency from `order`
Prior to this change, a full `Market` object was required to create an
`Order`. This would lead to an infinite regression if one wanted to
add an `Order` to a `Market` to keep track of any open orders that may
need to be canceled if the trade goes against us.

All the dependencies on a `market` object have been removed from `Order`.
In addition, the `OrderDetails` class has been enhanced to deal with
extra info required to place the order.