Remove trader-strategy-trader circulairy
Prior to this change, the stratgy class did instantiate a the trader
object by invoking App.get_running_app().trader from the strategy object
which was invoked from the trader object itself.

The strategy object is instntiated with a tradelock, which was the
object from the trader required by the strategy object.

Related to both #10 and #12.